refactor: . pull QueryRenderer methods up
