+++ src/Renderer/QueryRenderer.ts
@@ -1,19 +1,16 @@
 import type { EventRef, MarkdownPostProcessorContext } from 'obsidian';
 import { App, Keymap } from 'obsidian';
-import { GlobalFilter } from '../Config/GlobalFilter';
 import { GlobalQuery } from '../Config/GlobalQuery';
-import { PerformanceTracker } from '../lib/PerformanceTracker';
-import { explainResults, getQueryForQueryRenderer } from '../lib/QueryRendererHelper';
+import { getQueryForQueryRenderer } from '../lib/QueryRendererHelper';
 import type TasksPlugin from '../main';
 import { State } from '../Obsidian/Cache';
 import { getTaskLineAndFile, replaceTaskWithTasks } from '../Obsidian/File';
 import { TaskModal } from '../Obsidian/TaskModal';
 import type { TasksEvents } from '../Obsidian/TasksEvents';
-import type { QueryResult } from '../Query/QueryResult';
 import { TasksFile } from '../Scripting/TasksFile';
 import { DateFallback } from '../Task/DateFallback';
 import type { Task } from '../Task/Task';
-import { type QueryRendererParameters, QueryResultsRenderer } from './QueryResultsRenderer';
+import { QueryResultsRenderer } from './QueryResultsRenderer';
 import { createAndAppendElement } from './TaskLineRenderer';
 
 export class QueryRenderer {
@@ -138,80 +135,6 @@ class QueryRenderChild extends QueryResultsRenderer {
         }
 
         this.containerEl.firstChild?.replaceWith(content);
-    }
-
-    private async renderQuerySearchResults(
-        tasks: Task[],
-        state: State.Warm,
-        content: HTMLDivElement,
-        queryRendererParameters: QueryRendererParameters,
-    ) {
-        const queryResult = this.explainAndPerformSearch(state, tasks, content);
-
-        if (queryResult.searchErrorMessage !== undefined) {
-            // There was an error in the search, for example due to a problem custom function.
-            this.renderErrorMessage(content, queryResult.searchErrorMessage);
-            return;
-        }
-
-        await this.renderSearchResults(queryResult, content, queryRendererParameters);
-    }
-
-    private explainAndPerformSearch(state: State.Warm, tasks: Task[], content: HTMLDivElement) {
-        const measureSearch = new PerformanceTracker(`Search: ${this.query.queryId} - ${this.filePath}`);
-        measureSearch.start();
-
-        this.query.debug(`[render] Render called: plugin state: ${state}; searching ${tasks.length} tasks`);
-
-        if (this.query.queryLayoutOptions.explainQuery) {
-            this.createExplanation(content);
-        }
-
-        const queryResult = this.query.applyQueryToTasks(tasks);
-
-        measureSearch.finish();
-        return queryResult;
-    }
-
-    private async renderSearchResults(
-        queryResult: QueryResult,
-        content: HTMLDivElement,
-        queryRendererParameters: QueryRendererParameters,
-    ) {
-        const measureRender = new PerformanceTracker(`Render: ${this.query.queryId} - ${this.filePath}`);
-        measureRender.start();
-
-        await this.addAllTaskGroups(queryResult.taskGroups, content, queryRendererParameters);
-
-        const totalTasksCount = queryResult.totalTasksCount;
-        this.addTaskCount(content, queryResult);
-
-        this.query.debug(`[render] ${totalTasksCount} tasks displayed`);
-
-        measureRender.finish();
-    }
-
-    private renderErrorMessage(content: HTMLDivElement, errorMessage: string) {
-        content.createDiv().innerHTML = '<pre>' + `Tasks query: ${errorMessage.replace(/\n/g, '<br>')}` + '</pre>';
-    }
-
-    private renderLoadingMessage(content: HTMLDivElement) {
-        content.setText('Loading Tasks ...');
-    }
-
-    // Use the 'explain' instruction to enable this
-    private createExplanation(content: HTMLDivElement) {
-        const explanationAsString = explainResults(
-            this.source,
-            GlobalFilter.getInstance(),
-            GlobalQuery.getInstance(),
-            this.tasksFile,
-        );
-
-        const explanationsBlock = createAndAppendElement('pre', content);
-        explanationsBlock.addClasses(['plugin-tasks-query-explanation']);
-        explanationsBlock.setText(explanationAsString);
-        content.appendChild(explanationsBlock);
     }
 }
